refactor: extract item animation queue hook
